pulled supplement search into it's own component.
